requirements.txt
python-telegram-bot[job-queue]==21.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
APScheduler==3.10.4
